requirements.txt
flask
gunicorn==21.2.0
tensorflow-cpu
opencv-python
numpy
scipy